#237 [FIX] 수정 작업으로 인해 발생한 Miss Test Code 를, 통과 가능하게 수정
