+++ src/test/java/com/developlife/reviewtwits/project/ProjectApiTest.java
@@ -2,12 +2,9 @@ package com.developlife.reviewtwits.project;
 
 import com.developlife.reviewtwits.ApiTest;
 import com.developlife.reviewtwits.CommonDocument;
-import com.developlife.reviewtwits.entity.Project;
-import com.developlife.reviewtwits.entity.StatInfo;
 import com.developlife.reviewtwits.entity.User;
 import com.developlife.reviewtwits.message.request.user.RegisterUserRequest;
 import com.developlife.reviewtwits.repository.project.ProjectRepository;
-import com.developlife.reviewtwits.repository.project.StatInfoRepository;
 import com.developlife.reviewtwits.service.ProjectService;
 import com.developlife.reviewtwits.service.user.UserService;
 import com.developlife.reviewtwits.user.UserSteps;
@@ -20,12 +17,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-import org.springframework.jdbc.core.JdbcTemplate;
 
-import java.sql.Timestamp;
-import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.stream.IntStream;
 
 import static io.restassured.RestAssured.given;
@@ -46,11 +38,6 @@ public class ProjectApiTest extends ApiTest {
     ProjectService projectService;
     @Autowired
     ProjectRepository projectRepository;
-    @Autowired
-    StatInfoRepository statInfoRepository;
-
-    @Autowired
-    JdbcTemplate jdbcTemplate;
 
     private RegisterUserRequest registerUserRequest;
     private RegisterUserRequest registerAdminRequest;
@@ -145,390 +132,5 @@ public class ProjectApiTest extends ApiTest {
             .statusCode(HttpStatus.OK.value())
             .body("projectName", equalTo(request.projectName()))
             .log().all();
-    }
-
-    @Test
-    void 일간_방문_통계정보_검색성공_200() {
-        Project project = 통계_사전작업();
-        final String token = userSteps.로그인액세스토큰정보(UserSteps.로그인요청생성());
-
-        ExtractableResponse<Response> response = given(this.spec)
-                .filter(document(DEFAULT_RESTDOC_PATH, "일간 방문 통계정보를 검색하면, 올바른 입력값일 경우 200 OK 와 함께 정보가 반환됩니다.." +
-                                "<br>해당 유저가 가지고 있는 프로젝트의 아이디를 입력해야 합니다." +
-                                "<br>그리고 일간 방문 통계정보의 범위를 입력해야 합니다. " +
-                                "<br>범위로써 입력할 수 있는 정보는 아래와 같습니다." +
-                                "<br><br> 1d,3d,5d,7d,15d" +
-                                "<br> 1mo,3mo,6mo,1y,3y,5y" +
-                                "<br><br>위의 규칙에 맞지 않는 입력값일 경우, 400 Bad Request 가 반환됩니다." +
-                                "<br>헤더에 토큰 정보가 누락되었을 경우, 401 Unauthorized 가 반환됩니다." +
-                                "<br>해당 유저가 프로젝트를 소유하지 않을 경우, 403 Forbidden 이 반환됩니다." +
-                                "<br>입력받은 프로젝트 아이디로 된 프로젝트를 찾을 수 없을 경우, 404 Not Found 가 반환됩니다.", "일간방문통계그래프정보",
-                        CommonDocument.AccessTokenHeader,
-                        ProjectDocument.DailyVisitStatRequestParam,
-                        ProjectDocument.DailyVisitInfoResponseFields
-                        ))
-                .header("X-AUTH-TOKEN", token)
-                .param("projectId", project.getProjectId())
-                .param("range", ProjectSteps.exampleRange)
-                .when()
-                .get("/projects/daily-visit-graph-infos")
-                .then()
-                .assertThat()
-                .statusCode(HttpStatus.OK.value())
-                .log().all().extract();
-
-        JsonPath jsonPath = response.jsonPath();
-        assertThat(jsonPath.getString("range")).isEqualTo(ProjectSteps.exampleRange);
-        assertThat(jsonPath.getList("visitInfo.timeStamp")).size().isEqualTo(12);
-        assertThat(jsonPath.getList("visitInfo.visitCount")).size().isEqualTo(12);
-        assertThat(jsonPath.getList("visitInfo.previousCompare")).size().isEqualTo(12);
-    }
-
-    @Test
-    void 일간_방문_통계정보_헤더정보없음_401() {
-        Project project = 통계_사전작업();
-
-        given(this.spec)
-                .filter(document(DEFAULT_RESTDOC_PATH))
-                .param("projectId", project.getProjectId())
-                .param("range", ProjectSteps.exampleRange)
-                .when()
-                .get("/projects/daily-visit-graph-infos")
-                .then()
-                .assertThat()
-                .statusCode(HttpStatus.UNAUTHORIZED.value())
-                .log().all();
-    }
-
-    @Test
-    void 일간_방문_통계정보_접근권한_없음_403() {
-        Project project = 통계_사전작업();
-        final String token = userSteps.로그인액세스토큰정보(UserSteps.상대유저_로그인요청생성());
-
-        given(this.spec)
-                .filter(document(DEFAULT_RESTDOC_PATH, CommonDocument.ErrorResponseFields))
-                .header("X-AUTH-TOKEN", token)
-                .param("projectId", project.getProjectId())
-                .param("range", ProjectSteps.exampleRange)
-                .when()
-                .get("/projects/daily-visit-graph-infos")
-                .then()
-                .assertThat()
-                .statusCode(HttpStatus.FORBIDDEN.value())
-                .log().all();
-    }
-
-    @Test
-    void 일간_방문_통계정보_프로젝트아이디_없음_404() {
-        final String token = userSteps.로그인액세스토큰정보(UserSteps.로그인요청생성());
-
-        given(this.spec)
-                .filter(document(DEFAULT_RESTDOC_PATH, CommonDocument.ErrorResponseFields))
-                .header("X-AUTH-TOKEN", token)
-                .param("projectId", ProjectSteps.notExistProjectId)
-                .param("range", ProjectSteps.exampleRange)
-                .when()
-                .get("/projects/daily-visit-graph-infos")
-                .then()
-                .assertThat()
-                .statusCode(HttpStatus.NOT_FOUND.value())
-                .log().all().extract();
-    }
-
-    @Test
-    void 일간_방문_통계정보_프로젝트아이디_음수_400() {
-        final String token = userSteps.로그인액세스토큰정보(UserSteps.로그인요청생성());
-
-        given(this.spec)
-                .filter(document(DEFAULT_RESTDOC_PATH, CommonDocument.ErrorResponseFields))
-                .header("X-AUTH-TOKEN", token)
-                .param("projectId", ProjectSteps.wrongProjectId)
-                .param("range", ProjectSteps.exampleRange)
-                .when()
-                .get("/projects/daily-visit-graph-infos")
-                .then()
-                .assertThat()
-                .statusCode(HttpStatus.BAD_REQUEST.value())
-                .log().all().extract();
-    }
-    @Test
-    void 일간_방문_통계정보_검색시간구간_허용외값_400() {
-        Project project = 통계_사전작업();
-        final String token = userSteps.로그인액세스토큰정보(UserSteps.로그인요청생성());
-
-        given(this.spec)
-                .filter(document(DEFAULT_RESTDOC_PATH, CommonDocument.ErrorResponseFields))
-                .header("X-AUTH-TOKEN", token)
-                .param("projectId", project.getProjectId())
-                .param("range", ProjectSteps.wrongRange)
-                .when()
-                .get("/projects/daily-visit-graph-infos")
-                .then()
-                .assertThat()
-                .statusCode(HttpStatus.BAD_REQUEST.value())
-                .log().all().extract();
-    }
-
-    @Test
-    void 최근방문_통계정보_요청_성공_200(){
-        Project project = 통계_사전작업();
-        final String token = userSteps.로그인액세스토큰정보(UserSteps.로그인요청생성());
-
-        ExtractableResponse<Response> response = given(this.spec)
-                .filter(document(DEFAULT_RESTDOC_PATH, "최근 방문 통계정보를 검색하면, 올바른 입력값일 경우 200 OK 와 함께 정보가 반환됩니다.." +
-                                "<br>해당 유저가 가지고 있는 프로젝트의 아이디를 입력해야 합니다. 음수로 입력할 경우 400 Bad Request 가 반환됩니다." +
-                                "<br>헤더에 토큰 정보가 누락되었을 경우, 401 Unauthorized 가 반환됩니다." +
-                                "<br>해당 유저가 프로젝트를 소유하지 않을 경우, 403 Forbidden 이 반환됩니다." +
-                                "<br>입력받은 프로젝트 아이디로 된 프로젝트를 찾을 수 없을 경우, 404 Not Found 가 반환됩니다.",
-                        "최근방문통계정보요청",
-                        CommonDocument.AccessTokenHeader,
-                        ProjectDocument.ProjectIdRequestParam,
-                        ProjectDocument.RecentVisitStatResponseFields))
-                .header("X-AUTH-TOKEN", token)
-                .param("projectId", project.getProjectId())
-                .when()
-                .get("/projects/recent-visit-counts")
-                .then()
-                .assertThat()
-                .statusCode(HttpStatus.OK.value())
-                .log().all().extract();
-
-        JsonPath jsonPath = response.jsonPath();
-        assertThat(jsonPath.getInt("todayVisit")).isEqualTo(3);
-        assertThat(jsonPath.getInt("yesterdayVisit")).isEqualTo(2);
-        assertThat(jsonPath.getInt("totalVisit")).isEqualTo(20);
-    }
-
-    @Test
-    void 최근방문_통계정보_요청_헤더정보없음_401(){
-        Project project = 통계_사전작업();
-
-        given(this.spec)
-                .filter(document(DEFAULT_RESTDOC_PATH))
-                .param("projectId", project.getProjectId())
-                .when()
-                .get("/projects/recent-visit-counts")
-                .then()
-                .assertThat()
-                .statusCode(HttpStatus.UNAUTHORIZED.value())
-                .log().all();
-    }
-
-    @Test
-    void 최근방문_통계정보_요청_접근권한없음_403(){
-        Project project = 통계_사전작업();
-        final String token = userSteps.로그인액세스토큰정보(UserSteps.상대유저_로그인요청생성());
-
-        given(this.spec)
-                .filter(document(DEFAULT_RESTDOC_PATH, CommonDocument.ErrorResponseFields))
-                .header("X-AUTH-TOKEN", token)
-                .param("projectId", project.getProjectId())
-                .when()
-                .get("/projects/recent-visit-counts")
-                .then()
-                .assertThat()
-                .statusCode(HttpStatus.FORBIDDEN.value())
-                .log().all();
-    }
-
-    @Test
-    void 최근방문_통계정보_요청_프로젝트아이디_존재하지않음_404(){
-        final String token = userSteps.로그인액세스토큰정보(UserSteps.로그인요청생성());
-
-        given(this.spec)
-                .filter(document(DEFAULT_RESTDOC_PATH, CommonDocument.ErrorResponseFields))
-                .header("X-AUTH-TOKEN", token)
-                .param("projectId", ProjectSteps.notExistProjectId)
-                .when()
-                .get("/projects/recent-visit-counts")
-                .then()
-                .assertThat()
-                .statusCode(HttpStatus.NOT_FOUND.value())
-                .log().all();
-    }
-
-    @Test
-    void 최근방문_통계정보_요청_프로젝트아이디음수_400(){
-        final String token = userSteps.로그인액세스토큰정보(UserSteps.로그인요청생성());
-
-        given(this.spec)
-                .filter(document(DEFAULT_RESTDOC_PATH, CommonDocument.ErrorResponseFields))
-                .header("X-AUTH-TOKEN", token)
-                .param("projectId", ProjectSteps.wrongProjectId)
-                .when()
-                .get("/projects/recent-visit-counts")
-                .then()
-                .assertThat()
-                .statusCode(HttpStatus.BAD_REQUEST.value())
-                .log().all();
-    }
-
-    @Test
-    void 방문수_그래프_정보_요청_성공_200(){
-        Project project = 통계_사전작업();
-        final String token = userSteps.로그인액세스토큰정보(UserSteps.로그인요청생성());
-
-        ExtractableResponse<Response> response = given(this.spec)
-                .filter(document(DEFAULT_RESTDOC_PATH, "방문수 그래프 정보를 요청하면, 올바른 입력값일 경우 200 OK 와 함께 통걔 정보가 반환됩니다." +
-                                "<br>헤더에 토큰 정보가 누락되었을 경우, 401 Unauthorized 가 반환됩니다." +
-                                "<br>해당 유저가 프로젝트를 소유하지 않을 경우, 403 Forbidden 이 반환됩니다." +
-                                "<br>입력받은 프로젝트 아이디로 된 프로젝트를 찾을 수 없을 경우, 404 Not Found 가 반환됩니다." +
-                                "<br>해당 유저가 가지고 있는 프로젝트의 아이디를 입력해야 합니다." +
-                                "<br>또한 통계 범위와 구간을 입력해야 하며, 범위로써 입력할 수 있는 정보는 아래와 같습니다." +
-                                "<br><br> 1d,3d,5d,7d,15d" +
-                                "<br> 1mo,3mo,6mo,1y,3y,5y" +
-                                "<br><br>위의 규칙에 맞지 않는 입력값일 경우, 400 Bad Request 가 반환됩니다.",
-                        "방문수그래프정보요청",
-                        CommonDocument.AccessTokenHeader,
-                        ProjectDocument.VisitGraphInfoRequestParamFields,
-                        ProjectDocument.VisitGraphStatResponseFields))
-                .header("X-AUTH-TOKEN", token)
-                .param("projectId", project.getProjectId())
-                .param("interval", ProjectSteps.exampleInterval)
-                .param("range", ProjectSteps.exampleRange)
-                .when()
-                .get("/projects/visit-graph-infos")
-                .then()
-                .assertThat()
-                .statusCode(HttpStatus.OK.value())
-                .log().all().extract();
-
-        JsonPath jsonPath = response.jsonPath();
-        assertThat(jsonPath.getString("interval")).isEqualTo(ProjectSteps.exampleInterval);
-        assertThat(jsonPath.getString("range")).isEqualTo(ProjectSteps.exampleRange);
-        assertThat(jsonPath.getInt("presentVisit")).isEqualTo(3);
-        assertThat(jsonPath.getInt("previousVisit")).isEqualTo(2);
-        assertThat(jsonPath.getList("visitInfo.timeStamp")).isNotEmpty();
-        assertThat(jsonPath.getList("visitInfo.visitCount")).isNotEmpty();
-        assertThat(jsonPath.getList("visitInfo.previousCompare")).isNotEmpty();
-    }
-
-    @Test
-    void 방문수_그래프_정보_요청_헤더정보없음_401(){
-        Project project = 통계_사전작업();
-
-        given(this.spec)
-                .filter(document(DEFAULT_RESTDOC_PATH))
-                .param("projectId", project.getProjectId())
-                .param("interval", ProjectSteps.exampleInterval)
-                .param("range", ProjectSteps.exampleRange)
-                .when()
-                .get("/projects/visit-graph-infos")
-                .then()
-                .assertThat()
-                .statusCode(HttpStatus.UNAUTHORIZED.value())
-                .log().all();
-    }
-
-    @Test
-    void 방문수_그래프_정보_요청_접근권한없음_403(){
-        Project project = 통계_사전작업();
-        final String token = userSteps.로그인액세스토큰정보(UserSteps.상대유저_로그인요청생성());
-
-        given(this.spec)
-                .filter(document(DEFAULT_RESTDOC_PATH, CommonDocument.ErrorResponseFields))
-                .header("X-AUTH-TOKEN", token)
-                .param("projectId", project.getProjectId())
-                .param("interval", ProjectSteps.exampleInterval)
-                .param("range", ProjectSteps.exampleRange)
-                .when()
-                .get("/projects/visit-graph-infos")
-                .then()
-                .assertThat()
-                .statusCode(HttpStatus.FORBIDDEN.value())
-                .log().all();
-    }
-
-    @Test
-    void 방문수_그래프_정보_요청_프로젝트없음_404(){
-        final String token = userSteps.로그인액세스토큰정보(UserSteps.로그인요청생성());
-
-        given(this.spec)
-                .filter(document(DEFAULT_RESTDOC_PATH, CommonDocument.ErrorResponseFields))
-                .header("X-AUTH-TOKEN", token)
-                .param("projectId", ProjectSteps.notExistProjectId)
-                .param("interval", ProjectSteps.exampleInterval)
-                .param("range", ProjectSteps.exampleRange)
-                .when()
-                .get("/projects/visit-graph-infos")
-                .then()
-                .assertThat()
-                .statusCode(HttpStatus.NOT_FOUND.value())
-                .log().all();
-    }
-
-    @Test
-    void 방문수_그래프_정보_요청_프로젝트아이디_음수_400(){
-        final String token = userSteps.로그인액세스토큰정보(UserSteps.로그인요청생성());
-
-        given(this.spec)
-                .filter(document(DEFAULT_RESTDOC_PATH, CommonDocument.ErrorResponseFields))
-                .header("X-AUTH-TOKEN", token)
-                .param("projectId", ProjectSteps.wrongProjectId)
-                .param("interval", ProjectSteps.exampleInterval)
-                .param("range", ProjectSteps.exampleRange)
-                .when()
-                .get("/projects/visit-graph-infos")
-                .then()
-                .assertThat()
-                .statusCode(HttpStatus.BAD_REQUEST.value())
-                .log().all();
-    }
-
-    @Test
-    void 방문수_그래프_정보_요청_검색시간구간_허용외값_400(){
-        Project project = 통계_사전작업();
-        final String token = userSteps.로그인액세스토큰정보(UserSteps.로그인요청생성());
-
-        given(this.spec)
-                .filter(document(DEFAULT_RESTDOC_PATH, CommonDocument.ErrorResponseFields))
-                .header("X-AUTH-TOKEN", token)
-                .param("projectId", project.getProjectId())
-                .param("interval", ProjectSteps.exampleInterval)
-                .param("range", ProjectSteps.wrongRange)
-                .when()
-                .get("/projects/visit-graph-infos")
-                .then()
-                .assertThat()
-                .statusCode(HttpStatus.BAD_REQUEST.value())
-                .log().all();
-    }
-
-    Project 통계_사전작업() {
-        Project existedProject = projectRepository.findAll().get(0);
-
-        ArrayList<StatInfo> statInfos = new ArrayList<>();
-        for (int day = 1; day <= 10; day++) {
-            statInfos.add(ProjectSteps.통계정보_생성(existedProject, 2023, 3, day, 1));
-        }
-        for (int day = 1; day <= 5; day++) {
-            statInfos.add(ProjectSteps.통계정보_생성(existedProject, 2023, 3, day, 2));
-        }
-
-        LocalDateTime now = LocalDateTime.now();
-        int testYear = now.getYear();
-        int testMonth = now.getMonthValue();
-        int testDay = now.getDayOfMonth();
-        for (int hour = 1; hour <= 3; hour++) {
-            statInfos.add(ProjectSteps.통계정보_생성(existedProject, testYear, testMonth, testDay, hour));
-        }
-
-        for (int hour = 1; hour <= 2; hour++) {
-            statInfos.add(ProjectSteps.통계정보_생성(existedProject, testYear, testMonth, testDay - 1, hour));
-        }
-        saveAll(statInfos);
-        return existedProject;
-    }
-
-    void saveAll(List<StatInfo> statInfos) {
-        int index = 1;
-        for (StatInfo statInfo : statInfos) {
-            jdbcTemplate.update("insert into stat_info (stat_id,project_project_id,created_date) values (?,?,?)",
-                    index,
-                    statInfo.getProject().getProjectId(),
-                    Timestamp.valueOf(statInfo.getCreatedDate()));
-
-            index++;
-        }
     }
 }
